requirements.txt
pytest==6.2.5
wheel==0.36.2
click==8.0.3
my-stopwatch==1.0.1